requirements.txt
google-generativeai>=0.8.0
google-api-core>=2.15.0
openai>=1.12.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
pyright==1.1.407
nodeenv>=1.9.1
typing-extensions>=4.9.0
